% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderly.R
\name{orderly_remote_sharepoint}
\alias{orderly_remote_sharepoint}
\title{Create an orderly remote based on sharepoint}
\usage{
orderly_remote_sharepoint(url, site, path)
}
\arguments{
\item{url}{Sharepoint URL}

\item{site}{Sharepoint "site"}

\item{path}{Path within the sharepoint site}
}
\value{
An \code{orderly_remote_sharepoint} object
}
\description{
Create an orderly remote based on sharepoint
}
